Extract actions from bot index.
